+++ changelog.md
@@ -5,7 +5,7 @@ Uiua is not yet stable.
 ## 0.10.0 - 2024-??-??
 This version is not yet released. If you are reading this on the website, then these changes are live here.
 ### Language
-- Switch functions now format to use `⟨⟩` brackets
+- Switch functions now format to use `⟨⟩` brackets ⟨⟩
 - Switch functions now work with [`under ⍜`](https://uiua.org/docs/under)
 - [`under ⍜`](https://uiua.org/docs/under) [`join ⊂`](https://uiua.org/docs/join) now works with arrays of the same rank as long as the row count does not change